Fix level2/source.c

+++ level2/source.c
@@ -1,4 +1,6 @@
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 char 		*p(void)
 {
@@ -8,7 +10,8 @@ char 		*p(void)
 	fflush(stdout);
 	gets(buffer);
 	ret_addr = __builtin_return_address(0);
-	if (ret_addr & 0xb0000000 == 0xb0000000)
+	
+	if (((unsigned long)ret_addr & 0xb0000000) == 0xb0000000)
 	{
 		printf("(%p)\n", ret_addr);
 		exit(1);
